work on modularising install script

+++ install/emonSD_build_test.sh
@@ -16,10 +16,15 @@
 
 #! /bin/sh
 
+# auto detect install location (e.g /usr/emoncms or /home/pi)
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+usrdir=${DIR/\/emonpi\/install/}
+
 USER=pi
-homedir=/home/pi
 DEFAULT_SSH_PASSWORD=emonpi2016
 hostname=emonpi
+emonSD_pi_env=1
+mysql_password="emonpiemoncmsmysql2016"
 
 sudo apt-get update -y
 # sudo apt-get upgrade -y
@@ -31,15 +36,16 @@ sudo apt --fix-broken install
 
 # Emoncms install process from:
 # https://github.com/emoncms/emoncms/blob/master/docs/RaspberryPi/readme.md
-sudo apt-get install -y apache2 mariadb-server mysql-client php7.0 libapache2-mod-php7.0 php7.0-mysql php7.0-gd php7.0-opcache php7.0-curl php-pear php7.0-dev php7.0-mcrypt php7.0-common redis-server git build-essential php7.0-mbstring libmosquitto-dev mosquitto python-pip python-dev gettext ufw
+sudo apt-get install -y apache2 mariadb-server mysql-client php7.0 libapache2-mod-php7.0 php7.0-mysql php7.0-gd php7.0-opcache php7.0-curl php-pear php7.0-dev php7.0-mcrypt php7.0-common git build-essential php7.0-mbstring python-pip python-dev gettext
 
 # Install the pecl dependencies
 sudo pecl channel-update pecl.php.net
-printf "\n" | sudo pecl install redis Mosquitto-beta
 
-# --------------------------------------------------------------------------------
-# Redis configuration
-# --------------------------------------------------------------------------------
+echo "-------------------------------------------------------------"
+echo "Redis configuration"
+echo "-------------------------------------------------------------"
+sudo apt-get install -y redis-server
+printf "\n" | sudo pecl install redis
 
 # Add redis to php mods available 
 printf "extension=redis.so" | sudo tee /etc/php/7.0/mods-available/redis.ini 1>&2
@@ -51,9 +57,11 @@ sudo sed -i "s/^save 300 1/#save 300 1/" /etc/redis/redis.conf
 sudo sed -i "s/^save 60 1/#save 60 1/" /etc/redis/redis.conf
 sudo service redis-server restart
 
-# --------------------------------------------------------------------------------
-# Mosquitto configuration
-# --------------------------------------------------------------------------------
+echo "-------------------------------------------------------------"
+echo "Mosquitto configuration"
+echo "-------------------------------------------------------------"
+sudo apt-get install -y libmosquitto-dev mosquitto
+printf "\n" | sudo pecl install Mosquitto-beta
 
 # Add mosquitto to php mods available
 printf "extension=mosquitto.so" | sudo tee /etc/php/7.0/mods-available/mosquitto.ini 1>&2
@@ -72,13 +80,12 @@ sudo sed -i -n '/log_type error/!p;$a log_type error' /etc/mosquitto/mosquitto.c
 sudo touch /etc/mosquitto/passwd
 sudo mosquitto_passwd -b /etc/mosquitto/passwd emonpi emonpimqtt2016
 
-# --------------------------------------------------------------------------------
-# Apache configuration
-# --------------------------------------------------------------------------------
-
+echo "-------------------------------------------------------------"
+echo "Apache configuration"
+echo "-------------------------------------------------------------"
 # Enable apache mod rewrite
 sudo a2enmod rewrite
-cat <<EOF >> $homedir/emoncms.conf
+cat <<EOF >> $usrdir/emoncms.conf
 <Directory /var/www/html/emoncms>
     Options FollowSymLinks
     AllowOverride All
@@ -87,7 +94,7 @@ cat <<EOF >> $homedir/emoncms.conf
     Allow from all
 </Directory>
 EOF
-sudo mv $homedir/emoncms.conf /etc/apache2/sites-available/emoncms.conf
+sudo mv $usrdir/emoncms.conf /etc/apache2/sites-available/emoncms.conf
 # Review is this line needed? if so check for existing entry
 # printf "ServerName localhost" | sudo tee -a /etc/apache2/apache2.conf 1>&2
 sudo a2ensite emoncms
@@ -98,32 +105,26 @@ sudo sed -i "s/^CustomLog/#CustomLog/" /etc/apache2/conf-available/other-vhosts-
 
 sudo service apache2 restart
 
-# --------------------------------------------------------------------------------
-# Setup the Mariadb server (MYSQL)
-# --------------------------------------------------------------------------------
+echo "-------------------------------------------------------------"
+echo "Setup the Mariadb server (MYSQL)"
+echo "-------------------------------------------------------------"
 # Secure mysql
+echo "- Secure MYSQL"
 sudo mysql -e "DELETE FROM mysql.user WHERE User='root' AND Host NOT IN ('localhost', '127.0.0.1', '::1'); DELETE FROM mysql.user WHERE User=''; DROP DATABASE IF EXISTS test; DELETE FROM mysql.db WHERE Db='test' OR Db='test\_%'; FLUSH PRIVILEGES;"
 # Create the emoncms database using utf8 character decoding:
+echo "- Create emoncms database"
 sudo mysql -e "CREATE DATABASE emoncms DEFAULT CHARACTER SET utf8;"
 # Add emoncms database, set user permissions
-sudo mysql -e "CREATE USER 'emoncms'@'localhost' IDENTIFIED BY 'emonpiemoncmsmysql2016'; GRANT ALL ON emoncms.* TO 'emoncms'@'localhost'; flush privileges;"
+echo "- Add emoncms user and assign to database"
+sudo mysql -e "CREATE USER 'emoncms'@'localhost' IDENTIFIED BY '$mysql_password'; GRANT ALL ON emoncms.* TO 'emoncms'@'localhost'; flush privileges;"
 
-# Create data repositories for emoncms feed engines:
-sudo mkdir /var/lib/phpfiwa
-sudo mkdir /var/lib/phpfina
-sudo mkdir /var/lib/phptimeseries
-sudo chown www-data:root /var/lib/phpfiwa
-sudo chown www-data:root /var/lib/phpfina
-sudo chown www-data:root /var/lib/phptimeseries
-
-# --------------------------------------------------------------------------------
-# Install Emoncms Core
-# --------------------------------------------------------------------------------
-
+echo "-------------------------------------------------------------"
+echo "Install Emoncms Core"
+echo "-------------------------------------------------------------"
 # Emoncms install
 # Give pi user ownership over /var/www/ folder
 sudo chown $USER /var/www
-cd /var/www && git clone -b stable https://github.com/emoncms/emoncms.git
+cd /var/www && git clone -b master https://github.com/emoncms/emoncms.git
 
 # Create logfile
 sudo touch /var/log/emoncms.log
@@ -132,6 +133,15 @@ sudo chmod 666 /var/log/emoncms.log
 # Configure emoncms database settings
 # Make a copy of default.settings.php and call it settings.php:
 cd /var/www/emoncms && cp default.emonpi.settings.php settings.php
+
+# Create data repositories for emoncms feed engines:
+echo " - Creating phpfina, phpiwa, phptimeseries emoncms feed directories"
+sudo mkdir /var/lib/phpfiwa
+sudo mkdir /var/lib/phpfina
+sudo mkdir /var/lib/phptimeseries
+sudo chown www-data:root /var/lib/phpfiwa
+sudo chown www-data:root /var/lib/phpfina
+sudo chown www-data:root /var/lib/phptimeseries
 
 # Temporary: Replace this with default settings file
 sed -i "s/\/home\/pi\/data\/phpfina\//\/var\/lib\/phpfina\//" settings.php
@@ -149,10 +159,9 @@ echo "<?php header('Location: ../emoncms'); ?>" > /home/pi/index.php
 sudo mv /home/pi/index.php /var/www/html/index.php
 sudo rm /var/www/html/index.html
 
-# --------------------------------------------------------------------------------
-# Install Emoncms Services
-# --------------------------------------------------------------------------------
-
+echo "-------------------------------------------------------------"
+echo "Install Emoncms Services"
+echo "-------------------------------------------------------------"
 # Emoncms MQTT
 sudo ln -s /var/www/emoncms/scripts/services/emoncms_mqtt/emoncms_mqtt.service /lib/systemd/system
 sudo systemctl enable emoncms_mqtt.service
@@ -167,10 +176,9 @@ sudo ln -s /var/www/emoncms/scripts/services/service-runner/service-runner.servi
 sudo systemctl enable service-runner.service
 sudo systemctl start service-runner.service
 
-# --------------------------------------------------------------------------------
-# Install Emoncms Modules
-# --------------------------------------------------------------------------------
-
+echo "-------------------------------------------------------------"
+echo "Install Emoncms Modules"
+echo "-------------------------------------------------------------"
 # Review default branch: e.g stable
 cd /var/www/emoncms/Modules
 git clone https://github.com/emoncms/config.git
@@ -181,7 +189,6 @@ git clone https://github.com/emoncms/app.git
 git clone https://github.com/emoncms/wifi.git
 
 # Install emoncms modules that do not reside in /var/www/emoncms/Modules
-
 sudo mkdir /usr/emoncms
 sudo chown $USER /usr/emoncms
 mkdir /usr/emoncms/modules
@@ -215,76 +222,62 @@ git clone https://github.com/emoncms/usefulscripts.git
 
 # Symlink emoncms module folders here...
 # Review consistent approach here
+if [ "$emonSD_pi_env" = "1" ]; then
+    sudo visudo -cf $usrdir/emonpi/emoncms-rebootbutton && \
+    sudo cp $usrdir/emonpi/emoncms-rebootbutton /etc/sudoers.d/
+    sudo chmod 0440 /etc/sudoers.d/emoncms-rebootbutton
+    echo "Install emonPi Emoncms admin reboot button sudoers entry"
+            
+    sudo visudo -cf $usrdir/emonpi/emonhub-sudoers && \
+    sudo cp $usrdir/emonpi/emonhub-sudoers /etc/sudoers.d/
+    sudo chmod 0440 /etc/sudoers.d/emonhub-sudoers
+    echo "emonhub service control sudoers entry installed"
 
-sudo visudo -cf $homedir/emonpi/emoncms-rebootbutton && \
-sudo cp $homedir/emonpi/emoncms-rebootbutton /etc/sudoers.d/
-sudo chmod 0440 /etc/sudoers.d/emoncms-rebootbutton
-echo "Install emonPi Emoncms admin reboot button sudoers entry"
-        
-sudo visudo -cf $homedir/emonpi/emonhub-sudoers && \
-sudo cp $homedir/emonpi/emonhub-sudoers /etc/sudoers.d/
-sudo chmod 0440 /etc/sudoers.d/emonhub-sudoers
-echo "emonhub service control sudoers entry installed"
+    sudo visudo -cf $usrdir/emonpi/wifi-sudoers && \
+    sudo cp $usrdir/emonpi/wifi-sudoers /etc/sudoers.d/
+    sudo chmod 0440 /etc/sudoers.d/wifi-sudoers
+    echo "wifi sudoers entry installed"
 
-sudo visudo -cf $homedir/emonpi/wifi-sudoers && \
-sudo cp $homedir/emonpi/wifi-sudoers /etc/sudoers.d/
-sudo chmod 0440 /etc/sudoers.d/wifi-sudoers
-echo "wifi sudoers entry installed"
-
-sudo visudo -cf $homedir/emonpi/emoncms-setup/emoncms-setup-sudoers && \
-sudo cp $homedir/emonpi/emoncms-setup/emoncms-setup-sudoers /etc/sudoers.d/
-sudo chmod 0440 /etc/sudoers.d/emoncms-setup-sudoers
-echo "Emoncms setup module sudoers entry installed"
+    sudo visudo -cf $usrdir/emonpi/emoncms-setup/emoncms-setup-sudoers && \
+    sudo cp $usrdir/emonpi/emoncms-setup/emoncms-setup-sudoers /etc/sudoers.d/
+    sudo chmod 0440 /etc/sudoers.d/emoncms-setup-sudoers
+    echo "Emoncms setup module sudoers entry installed"
+fi
 
 # --------------------------------------------------------------------------------
-# Install EmonHub
+# EmonHub
 # --------------------------------------------------------------------------------
-# RaspberryPi Serial configuration
-# disable Pi3 Bluetooth and restore UART0/ttyAMA0 over GPIOs 14 & 15;
-# Review should this be: dtoverlay=pi3-miniuart-bt?
-sudo sed -i -n '/dtoverlay=pi3-disable-bt/!p;$a dtoverlay=pi3-disable-bt' /boot/config.txt
-
-# We also need to stop the Bluetooth modem trying to use UART
-sudo systemctl disable hciuart
-
-# Remove console from /boot/cmdline.txt
-sudo sed -i "s/console=serial0,115200 //" /boot/cmdline.txt
-
-# stop and disable serial service??
-sudo systemctl stop serial-getty@ttyAMA0.service
-sudo systemctl disable serial-getty@ttyAMA0.service
-
-cd /home/pi
-git clone https://github.com/openenergymonitor/emonhub.git
-mkdir data
-sudo apt-get install -y python-serial python-configobj
-sudo pip install paho-mqtt requests
-cd /home/pi/emonhub
-sudo ./install.systemd
-sudo systemctl start emonhub.service
-
-# Temporary: replace with update to default settings file
-sed -i "s/loglevel = DEBUG/loglevel = WARNING/" /home/pi/data/emonhub.conf
+$usrdir/emonpi/install/emonhub.sh $emonSD_pi_env
 
 # --------------------------------------------------------------------------------
 # EmonPi repo
 # --------------------------------------------------------------------------------
-cd /home/pi/
-git clone https://github.com/openenergymonitor/emonpi.git
-git clone https://github.com/openenergymonitor/RFM2Pi
+cd $usrdir
 
+if [ -d $usrdir/emonpi ]; then
+    git clone https://github.com/openenergymonitor/emonpi.git
+fi
+
+if [ -d $usrdir/RFM2Pi ]; then
+    git clone https://github.com/openenergymonitor/RFM2Pi
+fi
 
 # --------------------------------------------------------------------------------
 # EmonPi LCD Support
 # --------------------------------------------------------------------------------
-sudo apt-get install -y python-smbus i2c-tools python-rpi.gpio python-gpiozero
-sudo pip install xmltodict
+if [ "$emonSD_pi_env" = "1" ]; then
+    sudo apt-get install -y python-smbus i2c-tools python-rpi.gpio python-gpiozero
+    sudo pip install xmltodict
 
-# Uncomment dtparam=i2c_arm=on
-sudo sed -i "s/^#dtparam=i2c_arm=on/dtparam=i2c_arm=on/" /boot/config.txt
-# Append line i2c-dev to /etc/modules
-sudo sed -i -n '/i2c-dev/!p;$a i2c-dev' /etc/modules
+    # Uncomment dtparam=i2c_arm=on
+    sudo sed -i "s/^#dtparam=i2c_arm=on/dtparam=i2c_arm=on/" /boot/config.txt
+    # Append line i2c-dev to /etc/modules
+    sudo sed -i -n '/i2c-dev/!p;$a i2c-dev' /etc/modules
+fi
 
+# --------------------------------------------------------------------------------
+#
+# --------------------------------------------------------------------------------
 # Enable service-runner update
 # emonpi update checks for image type and only runs with a valid image name file in the boot partition
 sudo touch /boot/emonSD-30Oct18
@@ -296,13 +289,15 @@ sudo touch /boot/emonSD-30Oct18
 # Install log2ram, so that logging is on RAM to reduce SD card wear.
 # Logs are written to disk every hour or at shutdown
 # --------------------------------------------------------------------------------
-# Review: @pb66 modifications, rotated logs are moved out of ram&sync cycle
-curl -Lo log2ram.tar.gz https://github.com/azlux/log2ram/archive/master.tar.gz
-tar xf log2ram.tar.gz
-cd log2ram-master
-chmod +x install.sh && sudo ./install.sh
-cd ..
-rm -r log2ram-master
+if [ "$emonSD_pi_env" = "1" ]; then
+    # Review: @pb66 modifications, rotated logs are moved out of ram&sync cycle
+    curl -Lo log2ram.tar.gz https://github.com/azlux/log2ram/archive/master.tar.gz
+    tar xf log2ram.tar.gz
+    cd log2ram-master
+    chmod +x install.sh && sudo ./install.sh
+    cd ..
+    rm -r log2ram-master
+fi
 
 # --------------------------------------------------------------------------------
 # Enable serial uploads with avrdude and autoreset
@@ -314,16 +309,20 @@ git clone https://github.com/openenergymonitor/avrdude-rpi.git ~/avrdude-rpi && 
 # Misc
 # --------------------------------------------------------------------------------
 # Review: provide configuration file for default password and hostname
+if [ "$emonSD_pi_env" = "1" ]; then
+    # Set default SSH password:
+    printf "raspberry\n$DEFAULT_SSH_PASSWORD\n$DEFAULT_SSH_PASSWORD" | passwd
 
-# Set default SSH password:
-printf "raspberry\n$DEFAULT_SSH_PASSWORD\n$DEFAULT_SSH_PASSWORD" | passwd
+    # Set hostname
+    sudo sed -i "s/raspberrypi/$hostname/g" /etc/hosts
+    printf $hostname | sudo tee /etc/hostname > /dev/null
+fi
 
-# Set hostname
-sudo sed -i "s/raspberrypi/$hostname/g" /etc/hosts
-printf $hostname | sudo tee /etc/hostname > /dev/null
-
-# Configure UFW firewall
+# --------------------------------------------------------------------------------
+# UFW firewall
+# --------------------------------------------------------------------------------
 # Review: reboot required before running:
+sudo apt-get install -y ufw
 # sudo ufw allow 80/tcp
 # sudo ufw allow 443/tcp (optional, HTTPS not present)
 # sudo ufw allow 22/tcp
@@ -351,4 +350,6 @@ sudo service rsyslog restart
 # Review automated install: Emoncms Language Support
 # sudo dpkg-reconfigure locales
 
-sudo reboot
+if [ "$emonSD_pi_env" = "1" ]; then
+    sudo reboot
+fi
